+++ R/fit.R
@@ -14,19 +14,19 @@
 
 add_starting_values <- function(dist, x) {
   if(dist$dist == "gamma"){
-    dist$start <- list(scale = var(x) / mean(x),
-                       shape = mean(x)^2 / var(x)^2)
+    dist$start <- list(scale = stats::var(x) / mean(x),
+                       shape = mean(x)^2 / stats::var(x)^2)
   } else if(dist$distr == "gompertz"){
-    fit <- vglm(x~1, gompertz)
+    fit <- vglm(x~1, VGAM::gompertz)
     dist$start <- list(shape = exp(unname(coef(fit)[2])),
                        scale = exp(unname(coef(fit)[1])) )
   } else if(dist$distr == "lgumbel"){
     dist$start <- list(location = mean(log(x)),
-                       scale = pi*sd(log(x))/sqrt(6))
+                       scale = pi*stats::sd(log(x))/sqrt(6))
   } else if(dist$distr == "llog" ){
-    dist$start <- list(shape = mean(log(x)), scale = pi*sd(log(x))/sqrt(3))
+    dist$start <- list(shape = mean(log(x)), scale = pi*stats::sd(log(x))/sqrt(3))
   } else if(dist$distr == "pareto"){
-    fit <- vglm(x~1, paretoff)
+    fit <- vglm(x~1, VGAM::paretoff)
     dist$start  <- list(shape = exp(unname(coef(fit))))
     dist$fix.arg<- list(scale = fit@extra$scale)
   }
@@ -37,12 +37,12 @@ fit_dist_uncensored <- function(data, left, weight, dist) {
 
   x <- data[[left]]
 
-  dist %<>% list(data = x, distr = ., method = "mle")
+  dist <- list(data = x, distr = dist, method = "mle")
 
   if(!is.null(weight))
     dist$weights <- data[[weight]]
 
-  dist %<>% add_starting_values(x)
+  dist <- add_starting_values(dist, x)
   do.call(fitdistrplus::fitdist, dist)
 }
 
@@ -51,12 +51,12 @@ fit_dist_censored <- function(data, left, right, weight, dist) {
   x <- rowMeans(data[c(left, right)], na.rm = TRUE)
 
   censdata <- data.frame(left = data[[left]], right = data[[right]])
-  dist %<>% list(censdata = censdata, distr = .)
+  dist <- list(censdata = censdata, distr = dist)
 
   if(!is.null(weight))
     dist$weights <- data[[weight]]
 
-  dist %<>% add_starting_values(x)
+  dist <- add_starting_values(dist, x)
   do.call(fitdistrplus::fitdistcens, dist)
 }
 
@@ -68,21 +68,6 @@ remove_errors <- function(dist_fit, name, silent) {
   dist_fit$result
 }
 
-#' Fit Distribution
-#'
-#' Fits a distribution to species sensitivity data.
-#'
-#' @inheritParams ssd_fit_dists
-#' @param dist A string of the distribution to fit.
-#'
-#' @return An object of class \code{\link[fitdistrplus]{fitdist}}.
-#' @seealso \code{\link{ssd_fit_dists}}
-#' @export
-#'
-#' @examples
-#' ssd_fit_dist(boron_data)
-#' data(fluazinam)
-#' ssd_fit_dist(fluazinam, left = "left", right = "right")
 ssd_fit_dist <- function(
   data, left = "Conc", right = left, weight = NULL, dist = "lnorm") {
 
@@ -108,7 +93,7 @@ ssd_fit_dist <- function(
 #'
 #' By default the 'lnorm', 'llog', 'gompertz', 'lgumbel', 'gamma' and 'weibull'
 #' distributions are fitted to the data.
-#' The ssd_fit_dist and \code{\link{ssd_fit_dist}} functions have also been
+#' The ssd_fit_dists function has also been
 #' tested with the 'pareto' distribution.
 #'
 #' If weight specifies a column in the data frame with positive integers,
@@ -131,23 +116,23 @@ ssd_fit_dist <- function(
 #' @references
 #' Liu, Y., Salibián-Barrera, M., Zamar, R.H., and Zidek, J.V. 2018. Using artificial censoring to improve extreme tail quantile estimates. Journal of the Royal Statistical Society: Series C (Applied Statistics).
 #'
-#' @seealso \code{\link{ssd_fit_dist}}
 #' @examples
 #' ssd_fit_dists(boron_data)
-#' data(fluazinam)
+#' data(fluazinam, package = "fitdistrplus")
 #' ssd_fit_dists(fluazinam, left = "left", right = "right")
 ssd_fit_dists <- function(
   data, left = "Conc", right = left, weight = NULL,
   dists = c("lnorm", "llog", "gompertz", "lgumbel", "gamma", "weibull"),
   silent = FALSE
 ) {
-  check_vector(dists, length = c(1,.Machine$integer.max), unique = TRUE, named = FALSE)
+  check_vector(dists, length = TRUE, unique = TRUE, named = FALSE)
   check_flag(silent)
 
   safe_fit_dist <- safely(ssd_fit_dist)
   names(dists) <- dists
-  dists %<>% map(safe_fit_dist, data = data, left = left, right = right, weight = weight)
-  dists %<>% imap(remove_errors, silent = silent)
+  dists <- lapply(dists, safe_fit_dist, data = data, left = left, right = right, weight = weight)
+  dists <- mapply(remove_errors, dists, names(dists), 
+                  MoreArgs = list(silent = silent), SIMPLIFY = FALSE)
   dists <- dists[!vapply(dists, is.null, TRUE)]
   if(!length(dists)) stop("all distributions failed to fit", call. = FALSE)
   if(left == right) {