+++ NEWS.md
@@ -1,8 +1,21 @@
+# ssdtools 0.0.3
+
+- Added ssdtools-manual vignette
+- Changed predict() and ssd_hc() nboot argument from 1001 to 1000
+- Added hc5_boron data object
+- No longer export ssd_fit_dist() as ssd_fit_dists() renders redundant
+- geom_hcintersect() now takes multiple values
+- More information in DESCRIPTION
+- Added CRAN badge
+- Removed dependencies: dplyr, magrittr, plyr, purrr
+- Moved from depends to imports: VGAM, fitdistrplus, graphics, ggplot, stats
+- Moved from imports to suggests: tibble
+
 # ssdtools 0.0.2
 
-* Added contributors
-* Added hex
+- Added contributors
+- Added hex
 
 # ssdtools 0.0.1
 
-* Initial Release
+- Initial Release